scala code begun

+++ create_db.sql
@@ -3,9 +3,9 @@ CREATE TABLE articles (
     musician BOOLEAN,
     scientist BOOLEAN,
     article_name VARCHAR(255) NOT NULL,
-    mean_word_length FLOAT,
-    mean_sentence_length FLOAT,
-    stddev_word_length FLOAT,
-    stddev_sentence_length FLOAT,
+    mean_word_length DOUBLE,
+    mean_sentence_length DOUBLE,
+    stddev_word_length DOUBLE,
+    stddev_sentence_length DOUBLE,
     article_contents TEXT
 )
